Develop merge

+++ Tim26-ISS/src/main/java/com/example/demo/repository/DocumentRepository.java
@@ -5,9 +5,9 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface DocumentRepository extends JpaRepository<Document, Long> {
+public interface DocumentRepository extends JpaRepository<Document, Integer> {
 
     List<Document> findAll();
 
-    Document findOneById(int id);
+    Document findOneById(Integer id);
 }
