+++ Tim26-ISS/src/main/java/com/example/demo/repository/PanicRepository.java
@@ -3,5 +3,5 @@ package com.example.demo.repository;
 import com.example.demo.model.Panic;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PanicRepository extends JpaRepository<Panic, Long> {
+public interface PanicRepository extends JpaRepository<Panic, Integer> {
 }
