+++ Tim26-ISS/src/main/java/com/example/demo/repository/UserActivationRepository.java
@@ -3,5 +3,5 @@ package com.example.demo.repository;
 import com.example.demo.model.UserActivation;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserActivationRepository extends JpaRepository<UserActivation, Long> {
+public interface UserActivationRepository extends JpaRepository<UserActivation, Integer> {
 }
