+++ Tim26-ISS/src/main/java/com/example/demo/repository/WorkingHourRepository.java
@@ -3,5 +3,5 @@ package com.example.demo.repository;
 import com.example.demo.model.WorkingHour;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface WorkingHourRepository extends JpaRepository<WorkingHour, Long> {
+public interface WorkingHourRepository extends JpaRepository<WorkingHour, Integer> {
 }
